requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
databases==0.7.0
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
